add source image type
