+++ backend/source/tests/test_queries.py
@@ -2,7 +2,7 @@ import pytest
 from django.core.handlers.wsgi import WSGIRequest
 from graphene.test import Client as GrapheneClient
 
-from source.models import Source
+from source.models import Source, SourceImage
 from user.models import ContributorGroup, User
 
 
@@ -206,3 +206,27 @@ def test_resolve_public_source_editing_anonymous(
     assert (
         result["data"]["source"] is None
     ), "Public sources should not be accessible in editing interface without contributor access"
+
+
+def test_resolve_source_image(
+        graphql_client: GrapheneClient, anonymous_request: WSGIRequest,
+        source: Source, source_image: SourceImage,
+):
+    query = """
+        query Source($id: ID!) {
+            source(id: $id) {
+                id
+                image {
+                    url
+                    altText
+                    caption
+                }
+            }
+        }
+    """
+    variables = {"id": str(source.pk)}
+    result = graphql_client.execute(
+        query, variables=variables, context=anonymous_request,
+    )
+    assert result['data']['source']['image']['altText'] == source_image.alt_text
+
