+++ backend/source/types/SourceImageType.py
@@ -1,0 +1,15 @@
+from graphene_django import DjangoObjectType
+from graphene import NonNull, ResolveInfo, String
+
+from source.models import SourceImage
+
+class SourceImageType(DjangoObjectType):
+    url = NonNull(String)
+
+    class Meta:
+        model = SourceImage
+        fields = ["id", "source", "alt_text", "caption"]
+
+    @staticmethod
+    def resolve_url(parent: SourceImage, info: ResolveInfo):
+        return ""
